+++ tests/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2025-05-20
 ### Added
 - Added ability to specify alias in lambda env dictionary in such format: '${alias_name}'. Can be combined with wildcard
+- Added tests for aurora cluster and aurora instance
 ### Changed
 - Removed prefix, suffix and deploy_target_bucket from `init_parameters` as required parameters. 
 Instead, they will be extracted from syndicate.yml config